+++ include/libft/internal/array_int.h
@@ -6,7 +6,7 @@
 /*   By: abrabant <abrabant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/21 12:32:39 by abrabant          #+#    #+#             */
-/*   Updated: 2021/03/24 21:27:39 by abrabant         ###   ########.fr       */
+/*   Updated: 2021/03/27 13:16:14 by abrabant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@
 
 typedef struct s_array
 {
-	int		size;
-	int		length;
+	size_t	size;
+	size_t	length;
 	void	**array;
 }	t_array_int;
