+++ src/array/ft_array_get_length.c
@@ -6,13 +6,13 @@
 /*   By: abrabant <abrabant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/24 00:10:12 by abrabant          #+#    #+#             */
-/*   Updated: 2021/03/24 00:11:47 by abrabant         ###   ########.fr       */
+/*   Updated: 2021/03/27 13:25:14 by abrabant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft/internal/array_int.h"
 
-int	ft_array_get_length(t_array_int *a)
+size_t	ft_array_get_length(t_array_int *a)
 {
 	return (a->length);
 }
