+++ src/array/ft_array_get_size.c
@@ -6,13 +6,13 @@
 /*   By: abrabant <abrabant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/24 00:22:08 by abrabant          #+#    #+#             */
-/*   Updated: 2021/03/24 00:24:29 by abrabant         ###   ########.fr       */
+/*   Updated: 2021/03/27 13:25:25 by abrabant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft/internal/array_int.h"
 
-int	ft_array_get_size(t_array_int *a)
+size_t	ft_array_get_size(t_array_int *a)
 {
 	return (a->size);
 }
