+++ src/array/ft_array_inter.c
@@ -6,7 +6,7 @@
 /*   By: abrabant <abrabant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/23 01:00:20 by abrabant          #+#    #+#             */
-/*   Updated: 2021/03/24 21:19:55 by abrabant         ###   ########.fr       */
+/*   Updated: 2021/03/27 13:20:20 by abrabant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 t_array	ft_array_inter(t_array a1, t_array a2, int (*cmp)(void *, void *))
 {
 	t_array_int	*new;
-	int			i;
+	size_t		i;
 
 	new = (t_array_int *)ft_array_new(((t_array_int *)a1)->size);
 	if (new == NULL)
