+++ src/array/ft_array_uniq.c
@@ -6,7 +6,7 @@
 /*   By: abrabant <abrabant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/23 14:45:11 by abrabant          #+#    #+#             */
-/*   Updated: 2021/03/23 14:51:01 by abrabant         ###   ########.fr       */
+/*   Updated: 2021/03/27 13:22:05 by abrabant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_array_uniq(t_array_int *a, int cmp(void *, void *))
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
